feat(statistiques): calcul des statistiques prm
